Remove imaginate layer
