+++ editor/src/messages/frontend/frontend_message.rs
@@ -8,7 +8,6 @@ use crate::messages::prelude::*;
 use crate::messages::tool::utility_types::HintData;
 
 use graphene::color::Color;
-use graphene::layers::imaginate_layer::{ImaginateBaseImage, ImaginateGenerationParameters, ImaginateMaskFillContent, ImaginateMaskPaintMode};
 use graphene::layers::text_layer::Font;
 use graphene::LayerId;
 
@@ -53,36 +52,36 @@ pub enum FrontendMessage {
 		#[serde(rename = "isDefault")]
 		is_default: bool,
 	},
-	TriggerImaginateCheckServerStatus {
-		hostname: String,
-	},
-	TriggerImaginateGenerate {
-		parameters: ImaginateGenerationParameters,
-		#[serde(rename = "baseImage")]
-		base_image: Option<ImaginateBaseImage>,
-		#[serde(rename = "maskImage")]
-		mask_image: Option<ImaginateBaseImage>,
-		#[serde(rename = "maskPaintMode")]
-		mask_paint_mode: ImaginateMaskPaintMode,
-		#[serde(rename = "maskBlurPx")]
-		mask_blur_px: u32,
-		#[serde(rename = "maskFillContent")]
-		mask_fill_content: ImaginateMaskFillContent,
-		hostname: String,
-		#[serde(rename = "refreshFrequency")]
-		refresh_frequency: f64,
-		#[serde(rename = "documentId")]
-		document_id: u64,
-		#[serde(rename = "layerPath")]
-		layer_path: Vec<LayerId>,
-	},
-	TriggerImaginateTerminate {
-		#[serde(rename = "documentId")]
-		document_id: u64,
-		#[serde(rename = "layerPath")]
-		layer_path: Vec<LayerId>,
-		hostname: String,
-	},
+	// TriggerImaginateCheckServerStatus {
+	// 	hostname: String,
+	// },
+	// TriggerImaginateGenerate {
+	// 	parameters: ImaginateGenerationParameters,
+	// 	#[serde(rename = "baseImage")]
+	// 	base_image: Option<ImaginateBaseImage>,
+	// 	#[serde(rename = "maskImage")]
+	// 	mask_image: Option<ImaginateBaseImage>,
+	// 	#[serde(rename = "maskPaintMode")]
+	// 	mask_paint_mode: ImaginateMaskPaintMode,
+	// 	#[serde(rename = "maskBlurPx")]
+	// 	mask_blur_px: u32,
+	// 	#[serde(rename = "maskFillContent")]
+	// 	mask_fill_content: ImaginateMaskFillContent,
+	// 	hostname: String,
+	// 	#[serde(rename = "refreshFrequency")]
+	// 	refresh_frequency: f64,
+	// 	#[serde(rename = "documentId")]
+	// 	document_id: u64,
+	// 	#[serde(rename = "layerPath")]
+	// 	layer_path: Vec<LayerId>,
+	// },
+	// TriggerImaginateTerminate {
+	// 	#[serde(rename = "documentId")]
+	// 	document_id: u64,
+	// 	#[serde(rename = "layerPath")]
+	// 	layer_path: Vec<LayerId>,
+	// 	hostname: String,
+	// },
 	TriggerImport,
 	TriggerIndexedDbRemoveDocument {
 		#[serde(rename = "documentId")]
