+++ editor/src/messages/portfolio/document/properties_panel/properties_panel_message.rs
@@ -3,7 +3,6 @@ use crate::messages::layout::utility_types::widgets::assist_widgets::PivotPositi
 use crate::messages::portfolio::document::utility_types::misc::TargetDocument;
 use crate::messages::prelude::*;
 
-use graphene::layers::imaginate_layer::{ImaginateMaskFillContent, ImaginateMaskPaintMode, ImaginateSamplingMethod};
 use graphene::layers::style::{Fill, Stroke};
 use graphene::LayerId;
 
@@ -27,23 +26,6 @@ pub enum PropertiesPanelMessage {
 	ModifyTransform { value: f64, transform_op: TransformOp },
 	ResendActiveProperties,
 	SetActiveLayers { paths: Vec<Vec<LayerId>>, document: TargetDocument },
-	SetImaginateCfgScale { cfg_scale: f64 },
-	SetImaginateDenoisingStrength { denoising_strength: f64 },
-	SetImaginateLayerPath { layer_path: Option<Vec<LayerId>> },
-	SetImaginateMaskBlurPx { mask_blur_px: u32 },
-	SetImaginateMaskFillContent { mode: ImaginateMaskFillContent },
-	SetImaginateMaskPaintMode { paint: ImaginateMaskPaintMode },
-	SetImaginateNegativePrompt { negative_prompt: String },
-	SetImaginatePrompt { prompt: String },
-	SetImaginateRestoreFaces { restore_faces: bool },
-	SetImaginateSamples { samples: u32 },
-	SetImaginateSamplingMethod { method: ImaginateSamplingMethod },
-	SetImaginateScaleFromResolution,
-	SetImaginateSeed { seed: u64 },
-	SetImaginateSeedRandomize,
-	SetImaginateSeedRandomizeAndGenerate,
-	SetImaginateTiling { tiling: bool },
-	SetImaginateUseImg2Img { use_img2img: bool },
 	SetPivot { new_position: PivotPosition },
 	UpdateSelectedDocumentProperties,
 }
