+++ editor/src/messages/portfolio/document/properties_panel/utility_functions.rs
@@ -4,22 +4,20 @@ use crate::messages::layout::utility_types::layout_widget::{Layout, LayoutGroup,
 use crate::messages::layout::utility_types::misc::LayoutTarget;
 use crate::messages::layout::utility_types::widgets::assist_widgets::PivotAssist;
 use crate::messages::layout::utility_types::widgets::button_widgets::{IconButton, PopoverButton, TextButton};
-use crate::messages::layout::utility_types::widgets::input_widgets::{
-	CheckboxInput, ColorInput, DropdownEntryData, DropdownInput, FontInput, LayerReferenceInput, NumberInput, NumberInputMode, RadioEntryData, RadioInput, TextAreaInput, TextInput,
-};
+use crate::messages::layout::utility_types::widgets::input_widgets::{ColorInput, FontInput, NumberInput, NumberInputMode, RadioEntryData, RadioInput, TextAreaInput, TextInput};
 use crate::messages::layout::utility_types::widgets::label_widgets::{IconLabel, Separator, SeparatorDirection, SeparatorType, TextLabel};
-use crate::messages::portfolio::utility_types::{ImaginateServerStatus, PersistentData};
+use crate::messages::portfolio::utility_types::PersistentData;
 use crate::messages::prelude::*;
 
 use graphene::color::Color;
-use graphene::document::{pick_layer_safe_imaginate_resolution, Document};
-use graphene::layers::imaginate_layer::{ImaginateLayer, ImaginateMaskFillContent, ImaginateMaskPaintMode, ImaginateSamplingMethod, ImaginateStatus};
+use graphene::document::Document;
 use graphene::layers::layer_info::{Layer, LayerDataType, LayerDataTypeDiscriminant};
 use graphene::layers::nodegraph_layer::NodeGraphFrameLayer;
 use graphene::layers::style::{Fill, Gradient, GradientType, LineCap, LineJoin, Stroke};
 use graphene::layers::text_layer::{FontCache, TextLayer};
 
 use glam::{DAffine2, DVec2};
+use serde::__private::doc;
 use std::f64::consts::PI;
 use std::sync::Arc;
 
@@ -254,11 +252,6 @@ pub fn register_artwork_layer_properties(
 					tooltip: "Image".into(),
 					..Default::default()
 				})),
-				LayerDataType::Imaginate(_) => WidgetHolder::new(Widget::IconLabel(IconLabel {
-					icon: "NodeImaginate".into(),
-					tooltip: "Imaginate".into(),
-					..Default::default()
-				})),
 				LayerDataType::NodeGraphFrame(_) => WidgetHolder::new(Widget::IconLabel(IconLabel {
 					icon: "NodeNodes".into(),
 					tooltip: "Node Graph Frame".into(),
@@ -320,12 +313,6 @@ pub fn register_artwork_layer_properties(
 		LayerDataType::Image(_) => {
 			vec![node_section_transform(layer, persistent_data)]
 		}
-		LayerDataType::Imaginate(imaginate) => {
-			vec![
-				node_section_transform(layer, persistent_data),
-				node_section_imaginate(imaginate, layer, document, persistent_data, responses),
-			]
-		}
 		LayerDataType::NodeGraphFrame(node_graph_frame) => {
 			let is_graph_open = node_graph_message_handler.layer_path.as_ref().filter(|node_graph| *node_graph == &layer_path).is_some();
 			let selected_nodes = &node_graph_message_handler.selected_nodes;
@@ -335,7 +322,8 @@ pub fn register_artwork_layer_properties(
 				node_section_node_graph_frame(layer_path, node_graph_frame, is_graph_open),
 			];
 			if !selected_nodes.is_empty() && is_graph_open {
-				let parameters_sections = node_graph_message_handler.collate_properties(node_graph_frame);
+				let mut context = crate::messages::portfolio::document::node_graph::NodePropertiesContext { persistent_data, document, responses };
+				let parameters_sections = node_graph_message_handler.collate_properties(node_graph_frame, &mut context);
 				properties_sections.extend(parameters_sections.into_iter());
 			}
 			properties_sections
@@ -530,708 +518,6 @@ fn node_section_transform(layer: &Layer, persistent_data: &PersistentData) -> La
 			},
 		],
 	}
-}
-
-fn node_section_imaginate(imaginate_layer: &ImaginateLayer, layer: &Layer, document: &Document, persistent_data: &PersistentData, responses: &mut VecDeque<Message>) -> LayoutGroup {
-	let layer_reference_input_layer = imaginate_layer
-		.mask_layer_ref
-		.as_ref()
-		.and_then(|path| document.layer(path).ok())
-		.map(|layer| (layer.name.clone().unwrap_or_default(), LayerDataTypeDiscriminant::from(&layer.data)));
-
-	let layer_reference_input_layer_is_some = layer_reference_input_layer.is_some();
-
-	let layer_reference_input_layer_name = layer_reference_input_layer.as_ref().map(|(layer_name, _)| layer_name);
-	let layer_reference_input_layer_type = layer_reference_input_layer.as_ref().map(|(_, layer_type)| layer_type);
-
-	let mut layout = vec![
-		LayoutGroup::Row {
-			widgets: {
-				let tooltip = "Connection status to the server that computes generated images".to_string();
-
-				vec![
-					WidgetHolder::new(Widget::TextLabel(TextLabel {
-						value: "Server".into(),
-						tooltip: tooltip.clone(),
-						..Default::default()
-					})),
-					WidgetHolder::new(Widget::Separator(Separator {
-						separator_type: SeparatorType::Unrelated,
-						direction: SeparatorDirection::Horizontal,
-					})),
-					WidgetHolder::new(Widget::IconButton(IconButton {
-						size: 24,
-						icon: "Settings".into(),
-						tooltip: "Preferences: Imaginate".into(),
-						on_update: WidgetCallback::new(|_| DialogMessage::RequestPreferencesDialog.into()),
-						..Default::default()
-					})),
-					WidgetHolder::new(Widget::Separator(Separator {
-						separator_type: SeparatorType::Related,
-						direction: SeparatorDirection::Horizontal,
-					})),
-					WidgetHolder::new(Widget::TextLabel(TextLabel {
-						value: {
-							match &persistent_data.imaginate_server_status {
-								ImaginateServerStatus::Unknown => {
-									responses.push_back(PortfolioMessage::ImaginateCheckServerStatus.into());
-									"Checking...".into()
-								}
-								ImaginateServerStatus::Checking => "Checking...".into(),
-								ImaginateServerStatus::Unavailable => "Unavailable".into(),
-								ImaginateServerStatus::Connected => "Connected".into(),
-							}
-						},
-						bold: true,
-						tooltip,
-						..Default::default()
-					})),
-					WidgetHolder::new(Widget::Separator(Separator {
-						separator_type: SeparatorType::Related,
-						direction: SeparatorDirection::Horizontal,
-					})),
-					WidgetHolder::new(Widget::IconButton(IconButton {
-						size: 24,
-						icon: "Reload".into(),
-						tooltip: "Refresh connection status".into(),
-						on_update: WidgetCallback::new(|_| PortfolioMessage::ImaginateCheckServerStatus.into()),
-						..Default::default()
-					})),
-				]
-			},
-		},
-		LayoutGroup::Row {
-			widgets: {
-				let tooltip = "When generating, the percentage represents how many sampling steps have so far been processed out of the target number".to_string();
-
-				vec![
-					WidgetHolder::new(Widget::TextLabel(TextLabel {
-						value: "Progress".into(),
-						tooltip: tooltip.clone(),
-						..Default::default()
-					})),
-					WidgetHolder::new(Widget::Separator(Separator {
-						separator_type: SeparatorType::Unrelated,
-						direction: SeparatorDirection::Horizontal,
-					})),
-					WidgetHolder::new(Widget::TextLabel(TextLabel {
-						value: {
-							// Since we don't serialize the status, we need to derive from other state whether the Idle state is actually supposed to be the Terminated state
-							let mut interpreted_status = imaginate_layer.status.clone();
-							if imaginate_layer.status == ImaginateStatus::Idle && imaginate_layer.blob_url.is_some() && imaginate_layer.percent_complete > 0. && imaginate_layer.percent_complete < 100.
-							{
-								interpreted_status = ImaginateStatus::Terminated;
-							}
-
-							match interpreted_status {
-								ImaginateStatus::Idle => match imaginate_layer.blob_url {
-									Some(_) => "Done".into(),
-									None => "Ready".into(),
-								},
-								ImaginateStatus::Beginning => "Beginning...".into(),
-								ImaginateStatus::Uploading(percent) => format!("Uploading Base Image: {:.0}%", percent),
-								ImaginateStatus::Generating => format!("Generating: {:.0}%", imaginate_layer.percent_complete),
-								ImaginateStatus::Terminating => "Terminating...".into(),
-								ImaginateStatus::Terminated => format!("{:.0}% (Terminated)", imaginate_layer.percent_complete),
-							}
-						},
-						bold: true,
-						tooltip,
-						..Default::default()
-					})),
-				]
-			},
-		},
-		LayoutGroup::Row {
-			widgets: [
-				vec![
-					WidgetHolder::new(Widget::TextLabel(TextLabel {
-						value: "Image".into(),
-						tooltip: "Buttons that control the image generation process".into(),
-						..Default::default()
-					})),
-					WidgetHolder::new(Widget::Separator(Separator {
-						separator_type: SeparatorType::Unrelated,
-						direction: SeparatorDirection::Horizontal,
-					})),
-				],
-				{
-					match imaginate_layer.status {
-						ImaginateStatus::Beginning | ImaginateStatus::Uploading(_) => vec![WidgetHolder::new(Widget::TextButton(TextButton {
-							label: "Beginning...".into(),
-							tooltip: "Sending image generation request to the server".into(),
-							disabled: true,
-							..Default::default()
-						}))],
-						ImaginateStatus::Generating => vec![WidgetHolder::new(Widget::TextButton(TextButton {
-							label: "Terminate".into(),
-							tooltip: "Cancel the in-progress image generation and keep the latest progress".into(),
-							on_update: WidgetCallback::new(|_| DocumentMessage::ImaginateTerminate.into()),
-							..Default::default()
-						}))],
-						ImaginateStatus::Terminating => vec![WidgetHolder::new(Widget::TextButton(TextButton {
-							label: "Terminating...".into(),
-							tooltip: "Waiting on the final image generated after termination".into(),
-							disabled: true,
-							..Default::default()
-						}))],
-						ImaginateStatus::Idle | ImaginateStatus::Terminated => vec![
-							WidgetHolder::new(Widget::IconButton(IconButton {
-								size: 24,
-								icon: "Random".into(),
-								tooltip: "Generate with a new random seed".into(),
-								on_update: WidgetCallback::new(|_| PropertiesPanelMessage::SetImaginateSeedRandomizeAndGenerate.into()),
-								..Default::default()
-							})),
-							WidgetHolder::new(Widget::Separator(Separator {
-								separator_type: SeparatorType::Related,
-								direction: SeparatorDirection::Horizontal,
-							})),
-							WidgetHolder::new(Widget::TextButton(TextButton {
-								label: "Generate".into(),
-								tooltip: "Fill layer frame by generating a new image".into(),
-								on_update: WidgetCallback::new(|_| DocumentMessage::ImaginateGenerate.into()),
-								..Default::default()
-							})),
-							WidgetHolder::new(Widget::Separator(Separator {
-								separator_type: SeparatorType::Related,
-								direction: SeparatorDirection::Horizontal,
-							})),
-							WidgetHolder::new(Widget::TextButton(TextButton {
-								label: "Clear".into(),
-								tooltip: "Remove generated image from the layer frame".into(),
-								disabled: imaginate_layer.blob_url.is_none(),
-								on_update: WidgetCallback::new(|_| DocumentMessage::FrameClear.into()),
-								..Default::default()
-							})),
-						],
-					}
-				},
-			]
-			.concat(),
-		},
-		LayoutGroup::Row {
-			widgets: {
-				let tooltip = "Seed determines the random outcome, enabling limitless unique variations".to_string();
-
-				vec![
-					WidgetHolder::new(Widget::TextLabel(TextLabel {
-						value: "Seed".into(),
-						tooltip: tooltip.clone(),
-						..Default::default()
-					})),
-					WidgetHolder::new(Widget::Separator(Separator {
-						separator_type: SeparatorType::Unrelated,
-						direction: SeparatorDirection::Horizontal,
-					})),
-					WidgetHolder::new(Widget::IconButton(IconButton {
-						size: 24,
-						icon: "Regenerate".into(),
-						tooltip: "Set a new random seed".into(),
-						on_update: WidgetCallback::new(|_| PropertiesPanelMessage::SetImaginateSeedRandomize.into()),
-						..Default::default()
-					})),
-					WidgetHolder::new(Widget::Separator(Separator {
-						separator_type: SeparatorType::Related,
-						direction: SeparatorDirection::Horizontal,
-					})),
-					WidgetHolder::new(Widget::NumberInput(NumberInput {
-						value: Some(imaginate_layer.seed as f64),
-						min: Some(-1.),
-						tooltip,
-						on_update: WidgetCallback::new(move |number_input: &NumberInput| {
-							PropertiesPanelMessage::SetImaginateSeed {
-								seed: number_input.value.unwrap().round() as u64,
-							}
-							.into()
-						}),
-						..Default::default()
-					})),
-				]
-			},
-		},
-		LayoutGroup::Row {
-			widgets: {
-				let tooltip = "
-				Width and height of the image that will be generated. Larger resolutions take longer to compute.\n\
-				\n\
-				512x512 yields optimal results because the AI is trained to understand that scale best. Larger sizes may tend to integrate the prompt's subject more than once. Small sizes are often incoherent. Put the layer in a folder and resize that to keep resolution unchanged.\n\
-				\n\
-				Dimensions must be a multiple of 64, so these are set by rounding the layer dimensions. A resolution exceeding 1 megapixel is reduced below that limit because larger sizes may exceed available GPU memory on the server.
-				".trim().to_string();
-
-				vec![
-					WidgetHolder::new(Widget::TextLabel(TextLabel {
-						value: "Resolution".into(),
-						tooltip: tooltip.clone(),
-						..Default::default()
-					})),
-					WidgetHolder::new(Widget::Separator(Separator {
-						separator_type: SeparatorType::Unrelated,
-						direction: SeparatorDirection::Horizontal,
-					})),
-					WidgetHolder::new(Widget::IconButton(IconButton {
-						size: 24,
-						icon: "Rescale".into(),
-						tooltip: "Set the layer scale to this resolution".into(),
-						on_update: WidgetCallback::new(|_| PropertiesPanelMessage::SetImaginateScaleFromResolution.into()),
-						..Default::default()
-					})),
-					WidgetHolder::new(Widget::Separator(Separator {
-						separator_type: SeparatorType::Related,
-						direction: SeparatorDirection::Horizontal,
-					})),
-					WidgetHolder::new(Widget::TextLabel(TextLabel {
-						value: {
-							let (width, height) = pick_layer_safe_imaginate_resolution(layer, &persistent_data.font_cache);
-							format!("{} W x {} H", width, height)
-						},
-						tooltip,
-						bold: true,
-						..Default::default()
-					})),
-				]
-			},
-		},
-		LayoutGroup::Row {
-			widgets: {
-				let tooltip = "Number of iterations to improve the image generation quality, with diminishing returns around 40 when using the Euler A sampling method".to_string();
-				vec![
-					WidgetHolder::new(Widget::TextLabel(TextLabel {
-						value: "Sampling Steps".into(),
-						tooltip: tooltip.clone(),
-						..Default::default()
-					})),
-					WidgetHolder::new(Widget::Separator(Separator {
-						separator_type: SeparatorType::Unrelated,
-						direction: SeparatorDirection::Horizontal,
-					})),
-					WidgetHolder::new(Widget::NumberInput(NumberInput {
-						value: Some(imaginate_layer.samples.into()),
-						mode: NumberInputMode::Range,
-						range_min: Some(0.),
-						range_max: Some(150.),
-						is_integer: true,
-						min: Some(0.),
-						max: Some(150.),
-						tooltip,
-						on_update: WidgetCallback::new(move |number_input: &NumberInput| {
-							PropertiesPanelMessage::SetImaginateSamples {
-								samples: number_input.value.unwrap().round() as u32,
-							}
-							.into()
-						}),
-						..Default::default()
-					})),
-				]
-			},
-		},
-		LayoutGroup::Row {
-			widgets: {
-				let tooltip = "Algorithm used to generate the image during each sampling step".to_string();
-
-				let sampling_methods = ImaginateSamplingMethod::list();
-				let mut entries = Vec::with_capacity(sampling_methods.len());
-				for method in sampling_methods {
-					entries.push(DropdownEntryData {
-						label: method.to_string(),
-						on_update: WidgetCallback::new(move |_| PropertiesPanelMessage::SetImaginateSamplingMethod { method }.into()),
-						..DropdownEntryData::default()
-					});
-				}
-				let entries = vec![entries];
-
-				vec![
-					WidgetHolder::new(Widget::TextLabel(TextLabel {
-						value: "Sampling Method".into(),
-						tooltip: tooltip.clone(),
-						..Default::default()
-					})),
-					WidgetHolder::new(Widget::Separator(Separator {
-						separator_type: SeparatorType::Unrelated,
-						direction: SeparatorDirection::Horizontal,
-					})),
-					WidgetHolder::new(Widget::DropdownInput(DropdownInput {
-						entries,
-						selected_index: Some(imaginate_layer.sampling_method as u32),
-						tooltip,
-						..Default::default()
-					})),
-				]
-			},
-		},
-		LayoutGroup::Row {
-			widgets: {
-				let tooltip = "
-					Amplification of the text prompt's influence over the outcome. At 0, the prompt is entirely ignored.\n\
-					\n\
-					Lower values are more creative and exploratory. Higher values are more literal and uninspired, but may be lower quality.\n\
-					\n\
-					This parameter is otherwise known as CFG (classifier-free guidance).
-					"
-				.trim()
-				.to_string();
-
-				vec![
-					WidgetHolder::new(Widget::TextLabel(TextLabel {
-						value: "Text Guidance".into(),
-						tooltip: tooltip.to_string(),
-						..Default::default()
-					})),
-					WidgetHolder::new(Widget::Separator(Separator {
-						separator_type: SeparatorType::Unrelated,
-						direction: SeparatorDirection::Horizontal,
-					})),
-					WidgetHolder::new(Widget::NumberInput(NumberInput {
-						value: Some(imaginate_layer.cfg_scale),
-						mode: NumberInputMode::Range,
-						range_min: Some(0.),
-						range_max: Some(30.),
-						min: Some(0.),
-						max: Some(30.),
-						tooltip,
-						on_update: WidgetCallback::new(move |number_input: &NumberInput| {
-							PropertiesPanelMessage::SetImaginateCfgScale {
-								cfg_scale: number_input.value.unwrap(),
-							}
-							.into()
-						}),
-						..Default::default()
-					})),
-				]
-			},
-		},
-		LayoutGroup::Row {
-			widgets: vec![
-				WidgetHolder::new(Widget::TextLabel(TextLabel {
-					value: "Text Prompt".into(),
-					tooltip: "
-						Description of the desired image subject and style.\n\
-						\n\
-						Include an artist name like \"Rembrandt\" or art medium like \"watercolor\" or \"photography\" to influence the look. List multiple to meld styles.\n\
-						\n\
-						To boost (or lessen) the importance of a word or phrase, wrap it in parentheses ending with a colon and a multiplier, for example:\n\
-						\"Colorless green ideas (sleep:1.3) furiously\"
-						"
-					.trim()
-					.into(),
-					..Default::default()
-				})),
-				WidgetHolder::new(Widget::Separator(Separator {
-					separator_type: SeparatorType::Unrelated,
-					direction: SeparatorDirection::Horizontal,
-				})),
-				WidgetHolder::new(Widget::TextAreaInput(TextAreaInput {
-					value: imaginate_layer.prompt.clone(),
-					on_update: WidgetCallback::new(move |text_area_input: &TextAreaInput| {
-						PropertiesPanelMessage::SetImaginatePrompt {
-							prompt: text_area_input.value.clone(),
-						}
-						.into()
-					}),
-					..Default::default()
-				})),
-			],
-		},
-		LayoutGroup::Row {
-			widgets: vec![
-				WidgetHolder::new(Widget::TextLabel(TextLabel {
-					value: "Neg. Prompt".into(),
-					tooltip: "A negative text prompt can be used to list things like objects or colors to avoid".into(),
-					..Default::default()
-				})),
-				WidgetHolder::new(Widget::Separator(Separator {
-					separator_type: SeparatorType::Unrelated,
-					direction: SeparatorDirection::Horizontal,
-				})),
-				WidgetHolder::new(Widget::TextAreaInput(TextAreaInput {
-					value: imaginate_layer.negative_prompt.clone(),
-					on_update: WidgetCallback::new(move |text_area_input: &TextAreaInput| {
-						PropertiesPanelMessage::SetImaginateNegativePrompt {
-							negative_prompt: text_area_input.value.clone(),
-						}
-						.into()
-					}),
-					..Default::default()
-				})),
-			],
-		},
-		LayoutGroup::Row {
-			widgets: {
-				let tooltip = "Generate an image based upon the artwork beneath this frame in the containing folder".to_string();
-
-				vec![
-					WidgetHolder::new(Widget::TextLabel(TextLabel {
-						value: "Use Base Image".into(),
-						tooltip: tooltip.clone(),
-						..Default::default()
-					})),
-					WidgetHolder::new(Widget::Separator(Separator {
-						separator_type: SeparatorType::Unrelated,
-						direction: SeparatorDirection::Horizontal,
-					})),
-					WidgetHolder::new(Widget::CheckboxInput(CheckboxInput {
-						checked: imaginate_layer.use_img2img,
-						tooltip,
-						on_update: WidgetCallback::new(move |checkbox_input: &CheckboxInput| PropertiesPanelMessage::SetImaginateUseImg2Img { use_img2img: checkbox_input.checked }.into()),
-						..Default::default()
-					})),
-				]
-			},
-		},
-		LayoutGroup::Row {
-			widgets: {
-				let tooltip = "
-				Strength of the artistic liberties allowing changes from the base image. The image is unchanged at 0% and completely different at 100%.\n\
-				\n\
-				This parameter is otherwise known as denoising strength.
-				"
-				.trim()
-				.to_string();
-
-				vec![
-					WidgetHolder::new(Widget::TextLabel(TextLabel {
-						value: "Image Creativity".into(),
-						tooltip: tooltip.clone(),
-						..Default::default()
-					})),
-					WidgetHolder::new(Widget::Separator(Separator {
-						separator_type: SeparatorType::Unrelated,
-						direction: SeparatorDirection::Horizontal,
-					})),
-					WidgetHolder::new(Widget::NumberInput(NumberInput {
-						value: Some(imaginate_layer.denoising_strength * 100.),
-						unit: "%".into(),
-						mode: NumberInputMode::Range,
-						range_min: Some(0.),
-						range_max: Some(100.),
-						min: Some(0.),
-						max: Some(100.),
-						display_decimal_places: 2,
-						disabled: !imaginate_layer.use_img2img,
-						tooltip,
-						on_update: WidgetCallback::new(move |number_input: &NumberInput| {
-							PropertiesPanelMessage::SetImaginateDenoisingStrength {
-								denoising_strength: number_input.value.unwrap() / 100.,
-							}
-							.into()
-						}),
-						..Default::default()
-					})),
-				]
-			},
-		},
-		LayoutGroup::Row {
-			widgets: {
-				let tooltip = "
-				Reference to a layer or folder which masks parts of the base image. Image generation is constrained to masked areas.\n\
-				\n\
-				Black shapes represent the masked regions. Lighter shades of gray act as a partial mask, and colors become grayscale.
-				"
-				.trim()
-				.to_string();
-
-				vec![
-					WidgetHolder::new(Widget::TextLabel(TextLabel {
-						value: "Masking Layer".into(),
-						tooltip: tooltip.clone(),
-						..Default::default()
-					})),
-					WidgetHolder::new(Widget::Separator(Separator {
-						separator_type: SeparatorType::Unrelated,
-						direction: SeparatorDirection::Horizontal,
-					})),
-					WidgetHolder::new(Widget::LayerReferenceInput(LayerReferenceInput {
-						value: imaginate_layer.mask_layer_ref.clone(),
-						tooltip,
-						layer_name: layer_reference_input_layer_name.cloned(),
-						layer_type: layer_reference_input_layer_type.cloned(),
-						disabled: !imaginate_layer.use_img2img,
-						on_update: WidgetCallback::new(move |val: &LayerReferenceInput| PropertiesPanelMessage::SetImaginateLayerPath { layer_path: val.value.clone() }.into()),
-						..Default::default()
-					})),
-				]
-			},
-		},
-	];
-
-	if imaginate_layer.use_img2img && imaginate_layer.mask_layer_ref.is_some() && layer_reference_input_layer_is_some {
-		layout.extend(vec![
-			LayoutGroup::Row {
-				widgets: {
-					let tooltip = "
-					Constrain image generation to the interior (inpaint) or exterior (outpaint) of the mask, while referencing the other unchanged parts as context imagery.\n\
-					\n\
-					An unwanted part of an image can be replaced by drawing around it with a black shape and inpainting with that mask layer.\n\
-					\n\
-					An image can be uncropped by resizing the Imaginate layer to the target bounds and outpainting with a black rectangle mask matching the original image bounds.
-					"
-					.trim()
-					.to_string();
-
-					vec![
-						WidgetHolder::new(Widget::TextLabel(TextLabel {
-							value: "Mask Direction".to_string(),
-							tooltip: tooltip.clone(),
-							..Default::default()
-						})),
-						WidgetHolder::new(Widget::Separator(Separator {
-							separator_type: SeparatorType::Unrelated,
-							direction: SeparatorDirection::Horizontal,
-						})),
-						WidgetHolder::new(Widget::RadioInput(RadioInput {
-							entries: [(ImaginateMaskPaintMode::Inpaint, "Inpaint"), (ImaginateMaskPaintMode::Outpaint, "Outpaint")]
-								.into_iter()
-								.map(|(paint, name)| RadioEntryData {
-									label: name.to_string(),
-									on_update: WidgetCallback::new(move |_| PropertiesPanelMessage::SetImaginateMaskPaintMode { paint }.into()),
-									tooltip: tooltip.clone(),
-									..Default::default()
-								})
-								.collect(),
-							selected_index: imaginate_layer.mask_paint_mode as u32,
-							..Default::default()
-						})),
-					]
-				},
-			},
-			LayoutGroup::Row {
-				widgets: {
-					let tooltip = "Blur radius for the mask. Useful for softening sharp edges to blend the masked area with the rest of the image.".to_string();
-
-					vec![
-						WidgetHolder::new(Widget::TextLabel(TextLabel {
-							value: "Mask Blur".to_string(),
-							tooltip: tooltip.clone(),
-							..Default::default()
-						})),
-						WidgetHolder::new(Widget::Separator(Separator {
-							separator_type: SeparatorType::Unrelated,
-							direction: SeparatorDirection::Horizontal,
-						})),
-						WidgetHolder::new(Widget::NumberInput(NumberInput {
-							value: Some(imaginate_layer.mask_blur_px as f64),
-							unit: " px".into(),
-							mode: NumberInputMode::Range,
-							range_min: Some(0.),
-							range_max: Some(25.),
-							min: Some(0.),
-							is_integer: true,
-							tooltip,
-							on_update: WidgetCallback::new(move |number_input: &NumberInput| {
-								PropertiesPanelMessage::SetImaginateMaskBlurPx {
-									mask_blur_px: number_input.value.unwrap() as u32,
-								}
-								.into()
-							}),
-							..Default::default()
-						})),
-					]
-				},
-			},
-			LayoutGroup::Row {
-				widgets: {
-					let tooltip = "
-					Begin in/outpainting the masked areas using this fill content as the starting base image.\n\
-					\n\
-					Each option can be visualized by generating with 'Sampling Steps' set to 0.
-					"
-					.trim()
-					.to_string();
-
-					let mask_fill_content_modes = ImaginateMaskFillContent::list();
-					let mut entries = Vec::with_capacity(mask_fill_content_modes.len());
-					for mode in mask_fill_content_modes {
-						entries.push(DropdownEntryData {
-							label: mode.to_string(),
-							on_update: WidgetCallback::new(move |_| PropertiesPanelMessage::SetImaginateMaskFillContent { mode }.into()),
-							..DropdownEntryData::default()
-						});
-					}
-					let entries = vec![entries];
-
-					vec![
-						WidgetHolder::new(Widget::TextLabel(TextLabel {
-							value: "Mask Starting Fill".to_string(),
-							tooltip: tooltip.clone(),
-							..Default::default()
-						})),
-						WidgetHolder::new(Widget::Separator(Separator {
-							separator_type: SeparatorType::Unrelated,
-							direction: SeparatorDirection::Horizontal,
-						})),
-						WidgetHolder::new(Widget::DropdownInput(DropdownInput {
-							entries,
-							selected_index: Some(imaginate_layer.mask_fill_content as u32),
-							tooltip,
-							..Default::default()
-						})),
-					]
-				},
-			},
-		]);
-	}
-
-	layout.extend(vec![
-		LayoutGroup::Row {
-			widgets: {
-				let tooltip = "
-				Postprocess human (or human-like) faces to look subtly less distorted.\n\
-				\n\
-				This filter can be used on its own by enabling 'Use Base Image' and setting 'Sampling Steps' to 0.
-				"
-				.to_string();
-
-				vec![
-					WidgetHolder::new(Widget::TextLabel(TextLabel {
-						value: "Improve Faces".into(),
-						tooltip: tooltip.clone(),
-						..Default::default()
-					})),
-					WidgetHolder::new(Widget::Separator(Separator {
-						separator_type: SeparatorType::Unrelated,
-						direction: SeparatorDirection::Horizontal,
-					})),
-					WidgetHolder::new(Widget::CheckboxInput(CheckboxInput {
-						checked: imaginate_layer.restore_faces,
-						tooltip,
-						on_update: WidgetCallback::new(move |checkbox_input: &CheckboxInput| {
-							PropertiesPanelMessage::SetImaginateRestoreFaces {
-								restore_faces: checkbox_input.checked,
-							}
-							.into()
-						}),
-						..Default::default()
-					})),
-				]
-			},
-		},
-		LayoutGroup::Row {
-			widgets: {
-				let tooltip = "Generate the image so its edges loop seamlessly to make repeatable patterns or textures".to_string();
-
-				vec![
-					WidgetHolder::new(Widget::TextLabel(TextLabel {
-						value: "Tiling".into(),
-						tooltip: tooltip.clone(),
-						..Default::default()
-					})),
-					WidgetHolder::new(Widget::Separator(Separator {
-						separator_type: SeparatorType::Unrelated,
-						direction: SeparatorDirection::Horizontal,
-					})),
-					WidgetHolder::new(Widget::CheckboxInput(CheckboxInput {
-						checked: imaginate_layer.tiling,
-						tooltip,
-						on_update: WidgetCallback::new(move |checkbox_input: &CheckboxInput| PropertiesPanelMessage::SetImaginateTiling { tiling: checkbox_input.checked }.into()),
-						..Default::default()
-					})),
-				]
-			},
-		},
-	]);
-
-	LayoutGroup::Section { name: "Imaginate".into(), layout }
 }
 
 fn node_section_node_graph_frame(layer_path: Vec<graphene::LayerId>, node_graph_frame: &NodeGraphFrameLayer, open_graph: bool) -> LayoutGroup {
