+++ frontend/wasm/src/editor_api.rs
@@ -13,7 +13,6 @@ use editor::messages::input_mapper::utility_types::input_mouse::{EditorMouseStat
 use editor::messages::portfolio::utility_types::{ImaginateServerStatus, Platform};
 use editor::messages::prelude::*;
 use graphene::color::Color;
-use graphene::layers::imaginate_layer::ImaginateStatus;
 use graphene::LayerId;
 
 use serde::Serialize;
@@ -515,34 +514,34 @@ impl JsEditorHandle {
 
 	/// Sends the blob URL generated by JS to the Imaginate layer in the respective document
 	#[wasm_bindgen(js_name = setImaginateBlobURL)]
-	pub fn set_imaginate_blob_url(&self, document_id: u64, layer_path: Vec<LayerId>, blob_url: String, width: f64, height: f64) {
-		let resolution = (width, height);
-		let message = PortfolioMessage::ImaginateSetBlobUrl {
-			document_id,
-			layer_path,
-			blob_url,
-			resolution,
-		};
-		self.dispatch(message);
+	pub fn set_imaginate_blob_url(&self, _document_id: u64, _layer_path: Vec<LayerId>, _blob_url: String, _width: f64, _height: f64) {
+		// let resolution = (width, height);
+		// let message = PortfolioMessage::ImaginateSetBlobUrl {
+		// 	document_id,
+		// 	layer_path,
+		// 	blob_url,
+		// 	resolution,
+		// };
+		// self.dispatch(message);
 	}
 
 	/// Notifies the Imaginate layer of a new percentage of completion and whether or not it's currently generating
 	#[wasm_bindgen(js_name = setImaginateGeneratingStatus)]
-	pub fn set_imaginate_generating_status(&self, document_id: u64, path: Vec<LayerId>, percent: Option<f64>, status: String) {
-		let status = match status.as_str() {
-			"Idle" => ImaginateStatus::Idle,
-			"Beginning" => ImaginateStatus::Beginning,
-			"Uploading" => ImaginateStatus::Uploading(percent.expect("Percent needs to be supplied to set ImaginateStatus::Uploading")),
-			"Generating" => ImaginateStatus::Generating,
-			"Terminating" => ImaginateStatus::Terminating,
-			"Terminated" => ImaginateStatus::Terminated,
-			_ => panic!("Invalid string from JS for ImaginateStatus, received: {}", status),
-		};
+	pub fn set_imaginate_generating_status(&self, _document_id: u64, _path: Vec<LayerId>, _percent: Option<f64>, _status: String) {
+		// let status = match status.as_str() {
+		// 	"Idle" => ImaginateStatus::Idle,
+		// 	"Beginning" => ImaginateStatus::Beginning,
+		// 	"Uploading" => ImaginateStatus::Uploading(percent.expect("Percent needs to be supplied to set ImaginateStatus::Uploading")),
+		// 	"Generating" => ImaginateStatus::Generating,
+		// 	"Terminating" => ImaginateStatus::Terminating,
+		// 	"Terminated" => ImaginateStatus::Terminated,
+		// 	_ => panic!("Invalid string from JS for ImaginateStatus, received: {}", status),
+		// };
 
-		let percent = if matches!(status, ImaginateStatus::Uploading(_)) { None } else { percent };
+		// let percent = if matches!(status, ImaginateStatus::Uploading(_)) { None } else { percent };
 
-		let message = PortfolioMessage::ImaginateSetGeneratingStatus { document_id, path, percent, status };
-		self.dispatch(message);
+		// let message = PortfolioMessage::ImaginateSetGeneratingStatus { document_id, path, percent, status };
+		// self.dispatch(message);
 	}
 
 	/// Notifies the editor that the Imaginate server is available or unavailable
