+++ node-graph/graph-craft/src/lib.rs
@@ -5,6 +5,7 @@ pub mod document;
 pub mod proto;
 
 pub mod executor;
+pub mod imaginate_input;
 
 #[cfg(feature = "gpu")]
 pub mod gpu;
